+++ org/lazywizard/radar/util/ShapeUtils.java
@@ -40,7 +40,7 @@ public class ShapeUtils
         // Precalculate the sine and cosine
         // Instead of recalculating sin/cos for each line segment,
         // this algorithm rotates the line around the center point
-        final float theta = 2f * 3.1415926f / (float) numSegments;
+        final float theta = 2f * 3.1415926f / numSegments;
         final float cos = (float) FastTrig.cos(theta);
         final float sin = (float) FastTrig.sin(theta);
 
@@ -94,7 +94,7 @@ public class ShapeUtils
         // Precalculate the sine and cosine
         // Instead of recalculating sin/cos for each line segment,
         // this algorithm rotates the line around the center point
-        final float theta = 2f * 3.1415926f / (float) numSegments;
+        final float theta = 2f * 3.1415926f / numSegments;
         final float cos = (float) FastTrig.cos(theta);
         final float sin = (float) FastTrig.sin(theta);
         final float offsetCos = (float) FastTrig.cos(angleOffset);
@@ -155,7 +155,7 @@ public class ShapeUtils
         // Precalculate the sine and cosine
         // Instead of recalculating sin/cos for each line segment,
         // this algorithm rotates the line around the center point
-        final float theta = arcAngle / (float) (numSegments);
+        final float theta = arcAngle / numSegments;
         final float cos = (float) FastTrig.cos(theta);
         final float sin = (float) FastTrig.sin(theta);
